Fix rounding for reputation evidence averages

+++ supabase/migrations/20250322120000_public_profile_reputation_evidence.sql
@@ -150,8 +150,8 @@ SELECT
   CASE
     WHEN deal_metrics.first_deal_at IS NULL THEN 0
     ELSE ROUND(
-      deal_metrics.completed_deals_count
-        / NULLIF(DATE_PART('day', NOW() - deal_metrics.first_deal_at) / 30.0, 0),
+      (deal_metrics.completed_deals_count
+        / NULLIF(DATE_PART('day', NOW() - deal_metrics.first_deal_at) / 30.0, 0))::numeric,
       2
     )
   END AS avg_deals_per_month
